### 2021-09-27 #3767 @whedon
[PRE REVIEW]: sbp-env: A Python Package for Sampling-based Motion Planner and Samplers
**Submitting author:** @soraxas (<a href="http://orcid.org/0000-0003-0641-5250">Tin Lai</a>)
**Repository:** <a href="https://github.com/soraxas/sbp-env" target ="_blank">https://github.com/soraxas/sbp-env</a>
**Version:** v1.0.0
