### 2021-09-27 #3767 @soraxas
Hi @danielskatz thanks for the info!

Based on the given list, in reverse order: `KanishAnand, seungback, OlgerSiebinga, ekorudiawan, etc.` look like suitable candidates as reviewers (with *Robotics* as domain/topics and *python* as preferred language)
