+++ paper/master.bib
@@ -8,15 +8,16 @@
   doi          = {10/gdkx6g},
 }
 
-@online{jinmingwuCooperativePathfindingBased2019,
-  title         = {Cooperative {{Pathfinding}} Based on Memory-Efficient {{Multi}}-Agent {{RRT}}},
-  author        = {Jinmingwu, Jiang and Kaigui, Wu},
-  date          = {2019},
-  eprint        = {1911.03927},
-  eprinttype    = {arxiv},
-  archiveprefix = {arXiv},
+@article{jinmingwuCooperativePathfindingBased2019,
+  title={Cooperative Pathfinding based on memory-efficient Multi-agent RRT},
+  author={Jiang, Jinmingwu and Wu, Kaigui},
+  journal={IEEE Access},
+  volume={8},
+  pages={168743--168750},
+  year={2020},
+  publisher={IEEE},
+  doi= {10.1109/access.2020.3023200},
 }
-
 
 @inproceedings{klemmRRTConnectFaster2015,
   title      = {{{RRT}}∗-{{Connect}}: {{Faster}}, Asymptotically Optimal Motion Planning},
@@ -46,6 +47,7 @@
   date       = {2005},
   pages      = {3856--3861},
   publisher  = {{IEEE}},
+  doi = {10.1109/robot.2005.1570709}
 }
 
 
@@ -202,7 +204,8 @@
   booktitle    = {Proceedings of the 2001 IEEE International Symposium on Assembly and Task Planning (ISATP2001). Assembly and Disassembly in the Twenty-first Century.(Cat. No. 01TH8560)},
   pages        = {25--30},
   year         = {2001},
-  organization = {IEEE}
+  organization = {IEEE},
+  doi = {10.1109/isatp.2001.928961}
 }
 
 @article{sucan2012open,
@@ -233,6 +236,7 @@
   number    = {1},
   pages     = {18--19},
   year      = {2012},
-  publisher = {IEEE}
+  publisher = {IEEE},
+  doi = {10.1109/mra.2011.2181749}
 }
 
