### 2021-09-30 6430328
add contribution guide
Signed-off-by: Tin Lai <oscar@tinyiu.com>
