### 2021-09-30 2478bb5
add black formatting to ci
Signed-off-by: Tin Lai <oscar@tinyiu.com>
