+++ tests/test_planner_registry.py
@@ -20,7 +20,9 @@ class MySampler(Sampler):
 class TestRegisterPlanner(TestCase):
     def test_register(self):
         register_planner(
-            "my_id", planner_class=MyPlanner, sampler_id="my_sampler_id",
+            "my_id",
+            planner_class=MyPlanner,
+            sampler_id="my_sampler_id",
         )
 
     def test_register_not_derived(self):
@@ -33,33 +35,41 @@ class TestRegisterPlanner(TestCase):
 
     def test_register_duplicate_id(self):
         register_planner(
-            "my_duplicate_id", planner_class=MyPlanner, sampler_id="my_sampler_id",
+            "my_duplicate_id",
+            planner_class=MyPlanner,
+            sampler_id="my_sampler_id",
         )
         # cannot have duplicate id
         with self.assertRaises(ValueError):
             register_planner(
-                "my_duplicate_id", planner_class=MyPlanner, sampler_id="my_sampler_id",
+                "my_duplicate_id",
+                planner_class=MyPlanner,
+                sampler_id="my_sampler_id",
             )
 
 
 class TestRegisterSampler(TestCase):
     def test_register(self):
         register_sampler(
-            "my_sampler_id", sampler_class=MySampler,
+            "my_sampler_id",
+            sampler_class=MySampler,
         )
 
     def test_register_not_derived(self):
         with self.assertRaises(TypeError):
             register_sampler(
-                "my_non_derived_sampler_id", sampler_class=DummyClass,
+                "my_non_derived_sampler_id",
+                sampler_class=DummyClass,
             )
 
     def test_register_duplicate_id(self):
         register_sampler(
-            "my_duplicate_sampler_id", sampler_class=MySampler,
+            "my_duplicate_sampler_id",
+            sampler_class=MySampler,
         )
         # cannot have duplicate id
         with self.assertRaises(ValueError):
             register_sampler(
-                "my_duplicate_sampler_id", sampler_class=MySampler,
+                "my_duplicate_sampler_id",
+                sampler_class=MySampler,
             )
