### 2021-10-01 soraxas/sbp-env#4 @OlgerSiebinga
Exception after running the example from the documentation
When I run the example from the quick start page in the documentation, an exception occurs.

The command:
`python main.py rrt maps/room1.png`

The exception:
```
Traceback (most recent call last):
  File "main.py", line 287, in <module>
    environment.run()
  File "C:\Users\Olger\PycharmProjects\sbp-env\env.py", line 198, in run
    self.visualiser.terminates_hook()
  File "C:\Users\Olger\PycharmProjects\sbp-env\visualiser.py", line 148, in terminates_hook
    self.env_instance.sampler.visualiser.terminates_hook()
  File "C:\Users\Olger\PycharmProjects\sbp-env\env.py", line 126, in __getattr__
    return object.__getattribute__(self.visualiser, attr)
AttributeError: 'PygameEnvVisualiser' object has no attribute 'sampler'
```

The exception only occurs after the simulation has finished so it seems like a minor problem. Although I'm not really sure what happens at ` env.py, line 126, in __getattr__ ` and why. So, I don't have a proposed fix.
