### 2021-10-01 soraxas/sbp-env#5 @OlgerSiebinga
Suggestion to make installation easier
I was wondering why you have the following remark block in your installation instructions:
![image](https://user-images.githubusercontent.com/17888242/135623575-4d6acbc9-d3f0-479b-8e76-c1b0ba6aba98.png)

I think it would be easier to add those two packages to the file `requirements_klampt.txt`. That way they'll be installed automatically, it saves the user an extra action. Or is there any reason I'm missing why that shouldn't be done?
