### 2021-10-01 #3782 @OlgerSiebinga
Quick update from my side. I started reviewing today and I have opened two issues in the repository:

https://github.com/soraxas/sbp-env/issues/5#issue-1013318323
https://github.com/soraxas/sbp-env/issues/4#issue-1013302150

One is a suggestion for a minor change, the other is an exception that occurs. This exception does not seem hard to fix or important, but I'm not sure I understand it enough to propose a fix. Just posting the links here to have everything in one place.
