### 2021-10-02 1cc0756
Fix #4 and add corresponding tests
Signed-off-by: Tin Lai <oscar@tinyiu.com>
