### 2021-10-02 54e3486
improve instruction for missing start/goal
Signed-off-by: Tin Lai <oscar@tinyiu.com>
