### 2021-10-02 soraxas/sbp-env#5 @soraxas
The reason is because it is not a strict requirement for using the `klamp` backend (and it is also how `klampt` suggests their user for installation).

Opengl/pyqt are only used for visualisation, and they are not required for using `sbp-env` without visualisation. You can use that backend if you simply need to run it headlessly (e.g. benchmark on a server)

For example, if you invoke
```sh
python main.py rrt klampt_data/tx90blocks.xml --engine klampt
```
it will complain
```python
...
    from OpenGL import GL
ModuleNotFoundError: No module named 'OpenGL'
```

However, you can run experiment headlessly with
```sh
python main.py rrt klampt_data/tx90blocks.xml --engine klampt --no-display start 0,0,0,0,0,0 goal 1,1,1,1,1,1
```
which does not require opengl or pyqt
