### 2021-10-06 soraxas/sbp-env#6 @OlgerSiebinga
Graph building of prm planner without user information
The graph building method in the prm planner (`build_graph()` in `prmPlanner.py`) can take quite some time when a large number of nodes is used. However, the user is not notified that the planner is still processing data. The first time I encountered this, I suspected the software got stuck in an infinite loop because the window was not responding anymore. I think this can be easily fixed by adding a tqdm bar in the `build_graph()` method (at line 83)

(this suggestion is part of the JOSS review openjournals/joss-reviews#3782)
