### 2021-10-06 #3782 @OlgerSiebinga
Hi All,

Another update from my side: I've ticked a lot of boxes today and created two more issues in the repository. Besides those issues, I have two other (small) remarks.:

1. I think the statement of need is missing from the documentation, so that should be added
2. I feel that the paper would benefit from some editing for language. I think there are some tense, punctuation, and number agreement issues. However, I'm not a native English speaker nor an experienced editor, so I'm not sure if I'm the right person to help out with this. Do you have ideas on how to proceed?

Besides that, everything looks good!
