### 2021-10-06 #3782 @danielskatz
I'll go ahead and proof-read the paper now - I usually wait until the end because there can be changes made during the review process, but given your comment, maybe it's a good idea to do now
