+++ paper/paper.md
@@ -18,44 +18,44 @@ bibliography: master.bib
 
 # Background
 
-Sampling-based motion planning is one of the fundamental methods for robots to navigate and integrate with the real world [@elbanhawi2014_SampRobo].
+Sampling-based motion planning is one of the fundamental methods by which robots navigate and integrate with the real world [@elbanhawi2014_SampRobo].
 Motion planning involves planning the trajectories of the actuated part of the robot, under various constraints, while avoiding collisions with surrounding obstacles.
-Sampling-based motion planners (SBPs) are robust methods that avoid explicitly constructing the often intractable high-dimensional Configuration Space (C-Space).
-Instead, SBPs sample the C-Space randomly for valid connections and iteratively builds a roadmap of connectivity.
-Most SBPs are guaranteed to find a solution if one exists [@kavraki1996_AnalProb], and such a planner is said to be _probabilistic complete_ .
-Further developments on SBPs had granted _asymptotic optimality_[@elbanhawi2014_SampRobo]---a guarantee that the method will converge, in the limit, to the optimal solution.
+Sampling-based motion planners (SBPs) are robust methods that avoid explicitly constructing the often intractable high-dimensional configuration space (C-Space).
+Instead, SBPs randomly sample the C-Space for valid connections and iteratively build a roadmap of connectivity.
+Most SBPs are guaranteed to find a solution if one exists [@kavraki1996_AnalProb], and such a planner is said to be _probabilistic complete_.
+A further development for SBPs is _asymptotic optimality_[@elbanhawi2014_SampRobo]: a guarantee that the method will converge, in the limit, to the optimal solution.
 
-SBPs are applicable to a wide range of application
-For example, planning in arbitrary cost map [@iehlCostmapPlanningHigh2012], cooperative multi-agents planning [@jinmingwuCooperativePathfindingBased2019], as well as in dynamic environments [@yershova2005_DynaRRTs].
-On the one hand, researchers had focused on the algorithmic side of improving the graph or tree building [@lai2018_BalaGlob;@klemmRRTConnectFaster2015;@zhongTripleRrtsRobotPath2012;@elbanhawi2014_SampRobo;@lai2021lazyExperienceGraph;@lai2021rapidlyexploring].
+SBPs are applicable to a wide range of applications.
+Example include planning with arbitrary cost maps [@iehlCostmapPlanningHigh2012], cooperative multi-agent planning [@jinmingwuCooperativePathfindingBased2019], and planning in dynamic environments [@yershova2005_DynaRRTs].
+On the one hand, researchers have focused on the algorithmic side of improving the graph or tree building [@lai2018_BalaGlob;@klemmRRTConnectFaster2015;@zhongTripleRrtsRobotPath2012;@elbanhawi2014_SampRobo;@lai2021lazyExperienceGraph;@lai2021rapidlyexploring].
 On the other hand, the advancement of neural networks allows an abundance of learning approaches to be applied in SBPs [@strubAdaptivelyInformedTrees2020;@bagnell2014_ReinLear] and on improving the sampling distribution [@alcin2016_ExtrLear;@lai2020_BayeLoca;@lai2021plannerFlows;@laiLearningPlanOptimally2020;@lai2021diffSamp].
 
 
 # Statement of need
 
[…]
