### 2021-10-06 soraxas/sbp-env#9 @KanishAnand
Python version compatibility with scipy
Mentioning the requirement of python `version >= 3.8` in README would also help users the way it's done over [here](https://cs.tinyiu.com/sbp-env/quick-start.html#installation). Python `versions < 3.8` are not compatible with `scipy 1.6`

(Part of the JOSS review openjournals/joss-reviews#3782)
