### 2021-10-06 soraxas/sbp-env#10 @KanishAnand
Path recognition issue
I tried some source, destination positions with the following command and there seems some issue in recognition of the path.
`python main.py rrt maps/4d.png --engine 4d`

Attaching screenshot below:
![Screenshot from 2021-10-07 00-43-15](https://user-images.githubusercontent.com/43282223/136268396-2cacb234-6de7-4bb1-9be2-3c9274a03da5.png)

(Part of the JOSS review openjournals/joss-reviews#3782)
