### 2021-10-06 soraxas/sbp-env#10 @KanishAnand
Adding another example of the same issue upon running command
`python main.py rrt maps/room1.png --engine 4d --4d-robot-lengths 15,15`

![Screenshot from 2021-10-07 00-48-59](https://user-images.githubusercontent.com/43282223/136269411-5b313943-558f-4c78-b9a8-889ecb374c3f.png)
