### 2021-10-06 #3782 @danielskatz
> Hi @danielskatz
>
> Regarding checkpoint "_A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?_" I think more about the target audience is mentioned in the **Overview** section of the paper instead of **Statement of Need** section, so should that be an issue?

As long as the information is in the paper, I think it's ok.
