### 2021-10-07 soraxas/sbp-env#9 @soraxas
Hi I just tested the repo with virtualenv python 3.7, and I can confirm that python 3.7 should work.

The requirement is mentioned in README in the shield badge:
