### 2021-10-07 soraxas/sbp-env#9 @soraxas
Specifically I was testing it with `Python 3.7.10`
