### 2021-10-07 b51bdb2
add progress bar to prm graph building
Signed-off-by: Tin Lai <oscar@tinyiu.com>
