### 2021-10-07 soraxas/sbp-env#10 @KanishAnand
Yeah, they eventually got connected, and am clear with your explanation.
Thanks
