### 2021-10-07 d1493ab
add option to save stats output
Signed-off-by: Tin Lai <oscar@tinyiu.com>
