### 2021-10-07 soraxas/sbp-env#7 @soraxas
Thanks for the suggestion! Initially the stats are being displayed in the `tqdm` bar. The PR added new options to stream all results to a csv. You can now run
```sh
python main.py birrt maps/room1.png --save-output
```
and by default it will save a timestamped csv under `runs/` folder (output file is customisable via `--output-dir=STATS_DIR`)
