### 2021-10-11 soraxas/sbp-env#13 @KanishAnand
Test Instructions
Though it's standard, adding instruction to run tests in the documentation might be helpful for users wanting to contribute.

(Part of the JOSS review openjournals/joss-reviews#3782)
