### 2021-10-11 soraxas/sbp-env#7 @OlgerSiebinga
Great, this looks like a very nice solution. However, it does not seem to work as intended on my side and I'm not sure why. When I run the following command:

` python main.py rrt maps\room1.png start 100,100 goal 350,350 --save-output --output-dir=stats`

The result is an empty log file in a new folder called `runs`. So there are two issues, the `--output-dir` parameter seems to be ignored and the file is empty. The folder name could be easily fixed by sending it to the function `get_non_existing_filename`. But I cannot figure out why the written log file is empty. (A minor other issue is that the file has no extension, adding `.csv` would be good)

It might be relevant that I'm on Windows 10, if it works on Linux I can try to investigate why it does not work on Windows.

Finally, I was wondering if you are planning to add this functionality to the quick start part of the documentation. The documentation is complete as is, but an example of how to use this functionality would be very valuable in my opinion.

Let me know if I can help troubleshooting by trying out specific things/fixes.
