### 2021-10-13 e640b03
respect output-dir option when saving stats
Signed-off-by: Tin Lai <oscar@tinyiu.com>
