+++ env.py
@@ -183,7 +183,11 @@ class Env:
         self.started = True
 
         if self.args.save_output:
-            setup_csv_stats_logger(get_non_existing_filename())
+            setup_csv_stats_logger(
+                get_non_existing_filename(
+                    self.args.output_dir + "/%Y-%m-%d_%H:%M{}.csv"
+                )
+            )
             csv_logger = logging.getLogger("CSV_STATS")
             csv_logger.info(
                 [
