### 2021-10-14 soraxas/sbp-env#7 @soraxas
Thanks @OlgerSiebinga for your thorough details on the matter! I have found out that the empty logs in windows is due to the use of colon `:` character. It looks like the stats.csv works when I replace that invalid character. Documentation has also been added!
