### 2021-10-14 5a38116
modify test to be compliance on requirement of start and goal being valid
Signed-off-by: Tin Lai <oscar@tinyiu.com>
