+++ tests/common_vars.py
@@ -45,8 +45,8 @@ template_args = MagicDict(
     engine="image",
     epsilon=10.0,
     goalBias=0.02,
-    start_pt="0,0",
-    goal_pt="1,1",
+    start_pt="100,100",
+    goal_pt="350,350",
     num_dim=2,
     goal_radius=8.0,
     ignore_step_size=False,
