### 2021-10-15 cb3b9bd
update doi to long format
Signed-off-by: Tin Lai <oscar@tinyiu.com>
