+++ paper/master.bib
@@ -5,7 +5,7 @@
   journaltitle = {IEEE Access},
   volume       = {2},
   pages        = {56--77},
-  doi          = {10/gdkx6g},
+  doi          = {10.1109/access.2014.2302442},
 }
 
 @article{jinmingwuCooperativePathfindingBased2019,
@@ -26,7 +26,7 @@
   date       = {2015},
   pages      = {1670--1677},
   publisher  = {{IEEE}},
-  doi        = {10/ghpwwm},
+  doi        = {10.1109/robio.2015.7419012},
 }
 
 @inproceedings{zhongTripleRrtsRobotPath2012,
@@ -35,7 +35,7 @@
   author     = {Zhong, J. and Su, J.},
   date       = {2012-08},
   pages      = {188--192},
-  doi        = {10/ghjdnk},
+  doi        = {10.1109/csip.2012.6308826},
   eventtitle = {2012 {{International Conference}} on {{Computer Science}} and {{Information Processing}} ({{CSIP}})},
 }
 
@@ -58,7 +58,7 @@
   date       = {2012-07},
   pages      = {166--172},
   issn       = {2159-6255},
-  doi        = {10/gmv7jc},
+  doi        = {1010.1109/aim.2012.6265953},
   eventtitle = {2012 {{IEEE}}/{{ASME International Conference}} on {{Advanced Intelligent Mechatronics}} ({{AIM}})},
 }
 
@@ -70,7 +70,7 @@
   volume     = {4},
   pages      = {3020--3025},
   publisher  = {{IEEE}},
-  doi        = {10/d66p3t},
+  doi        = {10.1109/robot.1996.509171},
 }
 
 
@@ -112,7 +112,7 @@
   author     = {Lai, Tin and Ramos, Fabio and Francis, Gilad},
   date       = {2019},
   publisher  = {{IEEE}},
-  doi        = {10/ghqdw7},
+  doi        = {10.1109/icra.2019.8793618},
 }
 
 @inproceedings{strubAdaptivelyInformedTrees2020,
@@ -122,7 +122,7 @@
   date       = {2020-05},
   pages      = {3191--3198},
   issn       = {2577-087X},
-  doi        = {10/ghvk26},
+  doi        = {10.1109/icra40945.2020.9197338},
   eventtitle = {2020 {{IEEE International Conference}} on {{Robotics}} and {{Automation}} ({{ICRA}})},
 }
 
@@ -134,7 +134,7 @@
   volume = {97},
   pages = {9--67},
   issn = {9783642276446},
-  doi = {10/gddhk5},
+  doi = {10.1007/978-3-319-03194-1_2},
 }
 
 @article{alcin2016_ExtrLear,
@@ -145,7 +145,7 @@
   number       = {1},
   pages        = {1160--1163},
   issn         = {9781509018666},
-  doi          = {10/gddhmf},
+  doi          = {10.1109/mmar.2016.7575302},
 }
 
 
@@ -157,7 +157,7 @@
   volume       = {5},
   pages        = {1954--1961},
   publisher    = {{IEEE}},
-  doi          = {10/gg2n24},
+  doi          = {10.1109/lra.2020.2969145},
   number       = {2}
 }
 
