+++ paper/master.bib
@@ -58,7 +58,7 @@
   date       = {2012-07},
   pages      = {166--172},
   issn       = {2159-6255},
-  doi        = {1010.1109/aim.2012.6265953},
+  doi        = {10.1109/aim.2012.6265953},
   eventtitle = {2012 {{IEEE}}/{{ASME International Conference}} on {{Advanced Intelligent Mechatronics}} ({{AIM}})},
 }
 
