### 2021-10-15 #3782 @soraxas
Hi @danielskatz the DOIs should be all good to go
